refactor: share namespace filter helper across commands
