+++ cmd/micromize/root_test.go
@@ -64,51 +64,3 @@ func TestBuildDisabledSet(t *testing.T) {
 		})
 	}
 }
-
-func TestBuildNamespaceFilter(t *testing.T) {
-	tests := []struct {
-		name             string
-		filterNamespaces string
-		want             string
-	}{
-		{
-			name:             "empty filter returns only micromize exclusion",
-			filterNamespaces: "",
-			want:             "!micromize",
-		},
-		{
-			name:             "single namespace appends micromize exclusion",
-			filterNamespaces: "default",
-			want:             "!micromize,default",
-		},
-		{
-			name:             "multiple namespaces append micromize exclusion",
-			filterNamespaces: "default,kube-system",
-			want:             "!micromize,default,kube-system",
-		},
-		{
-			name:             "user already excludes micromize",
-			filterNamespaces: "default,!micromize",
-			want:             "!micromize,default",
-		},
-		{
-			name:             "only micromize exclusion",
-			filterNamespaces: "!micromize",
-			want:             "!micromize",
-		},
-		{
-			name:             "exclusion filter appends micromize exclusion",
-			filterNamespaces: "!kube-system",
-			want:             "!micromize,!kube-system",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := buildNamespaceFilter(tt.filterNamespaces)
-			if got != tt.want {
-				t.Errorf("buildNamespaceFilter(%q) = %q, want %q", tt.filterNamespaces, got, tt.want)
-			}
-		})
-	}
-}
